requirements.txt
dash>=2.17.0
pandas
pymongo
python-dotenv
